♻️ refactor(*): Renamed `lava/lavalink` to `lava/classes` and updated all references in the code
